Deleted old Image Carousel and renamed ImageCarouselNew
