+++ client/src/components/ImageCarousel.tsx
@@ -1,252 +1,201 @@
-import { useState, useEffect, useRef } from 'react';
-// import arrow from "../icons/s-arrow.png";
-import image1 from '../images/blue_frog.png';
-import image2 from '../images/banner.png';
-import image3 from '../images/tall_img.png';
+import { createContext, useContext, useEffect, useRef, useState, ReactNode } from 'react';
 import { ThemeIcon } from './ThemeIcon';
 
-//backend base url for getting images
-
-
-//TODO:
-//IMPORTANT: Remove this component? ImageCarouselNew.tsx is being called and seems to work as intended.
-//Finish Discover page version of this component
-
-//Default list of images to use
-const imageList: string[] = [image1, image2, image3];
-
-//Variable is identical in function to pages.
-//warnings about this variable can be found in Joseph Dunne's page documentation
-//All warnings shown about this variable in pages also applies to this one
-
-//This post was used to help create this component (found by Ben Gomez)
+// This post was used to help create this component (found by Ben Gomez)
 // https://blog.bitsrc.io/simple-carousel-in-react-2aac73887243
 
-//Will need to take in a list of images to render
-export const ImageCarousel = ({
-  carouselType,
-  dataList = imageList,
+const CarouselContext = createContext<{
+    // Index of the currently visible carousel item
+    currentIndex: number;
+    // Function to set the current index (wraps around boundaries)
+    handleIndexChange: (newIndex: number) => void;
+    // Function to notify when user hovers over carousel (pauses auto-scroll)
+    handleHover: (hovering: boolean) => void;
+    // List of React elements to display in the carousel
+    dataList: React.ReactNode[];
+}>({
+    currentIndex: 0,
+    handleIndexChange: () => {},
+    handleHover: () => {},
+    dataList: [],
+});
+
+/**
+ * Creates a button to navigate forward or backward in carousel order
+ * 
+ * @param direction 'left' or 'right' to indicate button direction
+ * @param className optional additional class names
+ * @param size 'small' for DiscoverCarousel, 'large' for ProjectCarousel 
+ * @returns A button to move the carousel in the direction specified
+ */
+export const CarouselButton = ({
+    direction,
+    className = '',
+    size = 'large',
 }: {
-  carouselType: 'Project' | 'Discover';
-  dataList?: string[];
+    direction: 'left' | 'right';
+    className?: string;
+    size: 'small' | 'large';
 }) => {
-  //State variable tracking what the currently displayed image's index is
-  const [currentIndex, setCurrentIndex] = useState(0);
-  //State variable tracking whether or not the user's mouse is hovering over this component
-  const [hovering, setHovering] = useState(false);
+    const { currentIndex, handleIndexChange } = useContext(CarouselContext);
+    // +1 or -1 based on direction
+    const directionNum = direction === 'left' ? -1 : 1;
 
-  const skipAuto = useRef(false);
+    return (
+        <button
+            className={`${className} carousel-btn-${direction}`}
+            onClick={() => handleIndexChange(currentIndex + directionNum)}
+        >
+            <ThemeIcon
+                id={size === 'small' ? 'dropdown-arrow' : 'carousel-arrow'}
+                width={size === 'small' ? 17 : 24}
+                height={size === 'small' ? 18 : 68}
+                className={'color-fill'}
+                ariaLabel={direction === 'left' ? 'Previous Image' : 'Next Image'}
+            />
+        </button>
+    );
+};
 
-  //Function to handle a change in which image index to use
-  const handleIndexChange = (newIndex: number) => {
-    //Skip next autoscroll due to button click
-    skipAuto.current = true;
-    if (newIndex > dataList.length - 1) {
-      //If new index is greater than max index, set index to 0
-      newIndex = 0;
-    } else if (newIndex < 0) {
-      //If new index is less than 0, set it to max index
-      newIndex = dataList.length - 1;
-    }
+/**
+ * Displays a series of buttons (tabs) to indicate current carousel item and allow navigation.
+ * 
+ * @param className optional additional class names
+ * @returns A JSX div element containing navigation buttons for the carousel
+ */
+export const CarouselTabs = ({ className = '' }: { className?: string }) => {
+    const { currentIndex, handleIndexChange, dataList } = useContext(CarouselContext);
 
-    //set new index
-    setCurrentIndex(newIndex);
-  };
-
-  const handleHover = (hovering: boolean) => {
-    if (hovering) {
-      skipAuto.current = true;
-      setHovering(true);
-    } else {
-      setHovering(false);
-    }
-  };
-
-  //Function called regularly to automatically scroll through carousel images
-  const autoScroll = () => {
-    //Is a condition met to skip the next autoscroll?
-    if (skipAuto.current) {
-      //If so, is it caused by hovering?
-      if (hovering) {
-        //If so, skip function as well as next function
-        return;
-      } else {
-        //If not, just skip this funciton
-        skipAuto.current = false;
-        return;
-      }
-    }
-
-    //Otherwise, run function as normal
-    if (currentIndex === dataList.length - 1) {
-      setCurrentIndex(0);
-    } else {
-      setCurrentIndex(currentIndex + 1);
-    }
-  };
-
-  useEffect(() => {
-    const interval = setInterval(() => {
-      autoScroll();
-    }, 10_000);
-    return () => clearInterval(interval);
-  });
-
-  const ProjectImageCarousel =
-    dataList != null ? (
-      <>
-        {
-          <>
-            {/* Main image content of carousel */}
-            <div
-              id="project-image-carousel-content"
-              onMouseEnter={() => handleHover(true)}
-              onMouseLeave={() => handleHover(false)}
-            >
-              {dataList.map((image, index) => {
+    return (
+        <div className={`carousel-tabs ${className}`}>
+            {dataList.map((_, index) => {
+                const active = index === currentIndex ? ' carousel-tab-active' : '';
                 return (
-                  <div
-                    className="project-image-carousel-item"
-                    key={index}
-                    style={{ transform: `translate(-${currentIndex * 100}%)` }}
-                  >
-                    <img
-                      className="project-image-carousel-image"
-                      src={`images/projects/${image}`}
-                      alt={'project image'}
-                      // Cannot use usePreloadedImage function because this is in a callback
-                      onLoad={(e) => {
-                        const projectImg = e.target as HTMLImageElement;
-                        projectImg.src = `images/projects/${image}`;
-                      }}
-                      onError={(e) => {
-                        const projectImg = e.target as HTMLImageElement;
-                        projectImg.src = image;
-                      }}
-                    />
-                  </div>
+                    <button
+                        className={`carousel-tab${active}`}
+                        onClick={() => handleIndexChange(index)}
+                        key={index}
+                    ></button>
                 );
-              })}
-            </div>
-
-            {/* Buttons for controlling carousel images */}
-            <div id="project-image-carousel-buttons">
-              <button
-                onClick={() => handleIndexChange(currentIndex - 1)}
-                id="project-image-carousel-left"
-              >
-                <ThemeIcon id={'carousel-arrow'} width={24} height={68} className={'color-fill'} ariaLabel={'Previous Image'}/>
-              </button>
-              <div id="carousel-tabs">
-                {dataList.map((_image, index) => {
-                  const className =
-                    index === currentIndex ? 'carousel-tab carousel-tab-active' : 'carousel-tab';
-                  return (
-                    <button
-                      className={className}
-                      onClick={() => handleIndexChange(index)}
-                      key={index}
-                    >
-                      <img />
-                    </button>
-                  );
-                })}
-              </div>
-              <button
-                onClick={() => handleIndexChange(currentIndex + 1)}
-                id="project-image-carousel-right"
-              >
-                <ThemeIcon id={'carousel-arrow'} width={24} height={68} className={'color-fill'} ariaLabel={'Next Image'}/>
-              </button>
-            </div>
-          </>
-        }
-      </>
-    ) : (
-      <>No images could be loaded</>
+            })}
+        </div>
     );
+};
 
-  const DiscoverCarousel =
-    dataList != undefined ? (
-      <>
-        {
-          <>
-            <div id="discover-carousel-bar">
-              <button
-                onClick={() => handleIndexChange(currentIndex - 1)}
-                id="discover-carousel-left"
-              >
-                <ThemeIcon id={'carousel-arrow'} width={24} height={68} className={'color-fill'} ariaLabel={'Previous Image'}/>
-              </button>
-              <div
-                id="discover-carousel-content"
-                onMouseEnter={() => handleHover(true)}
-                onMouseLeave={() => handleHover(false)}
-              >
-                {dataList.map((image, index) => (
-                  <div
-                    className="discover-carousel-item"
+// Only displays the current index of dataList (dataList MUST be an array of elements to work)
+/**
+ * Displays the current carousel item(s) and handles hover interactions.
+ * @param className optional additional class names
+ * @returns A JSX div element containing carousel content elements
+ */
+export const CarouselContent = ({ className = '' }: { className?: string }) => {
+    const { currentIndex, handleHover, dataList } = useContext(CarouselContext);
+
+    return (
+        <div className="carousel-contents">
+            {dataList.map((data, index) => (
+                <div
+                    className={className}
                     key={index}
+                    onMouseEnter={() => handleHover(true)}
+                    onMouseLeave={() => handleHover(false)}
                     style={{ transform: `translate(-${currentIndex * 100}%)` }}
-                  >
-                    <img className="discover-carousel-image" src={image} />
-                    <div className="discover-carousel-title">Title here</div>
-                    <div className="discover-carousel-description">
-                      Description here Lorem ipsum dolor sit amet consectetur adipisicing elit.
-                      Cumque, quidem? Fugit deleniti reiciendis doloremque assumenda neque rem
-                      repudiandae hic odit quaerat. Debitis dicta fugit expedita, quaerat autem
-                      animi quia voluptatum quam magni minus ducimus, aliquam obcaecati numquam.
-                      Quisquam similique laudantium doloremque pariatur expedita dicta suscipit
-                      eligendi molestiae, harum iusto sequi.
-                    </div>
-                  </div>
-                ))}
-              </div>
-
-              <button className="discover-carousel-link">
-                Learn More
-                <ThemeIcon id={'carousel-arrow'} width={24} height={68} className={'color-fill'} ariaLabel={'Next Image'}/>
-              </button>
-
-              <button
-                onClick={() => handleIndexChange(currentIndex + 1)}
-                id="discover-carousel-right"
-              >
-                <ThemeIcon id={'carousel-arrow'} width={24} height={68} className={'color-fill'} ariaLabel={'Next Image'}/>
-              </button>
-            </div>
-
-            <div id="discover-carousel-tabs">
-              <div id="carousel-tabs">
-                {dataList.map((_image, index) => {
-                  const className =
-                    index === currentIndex ? 'carousel-tab carousel-tab-active' : 'carousel-tab';
-                  return (
-                    <button
-                      className={className}
-                      onClick={() => handleIndexChange(index)}
-                      key={index}
-                    >
-                      <img />
-                    </button>
-                  );
-                })}
-              </div>
-            </div>
-          </>
-        }
-      </>
-    ) : (
-      <>Data failed to load</>
+                >
+                    {data}
+                </div>
+            ))}
+        </div>
     );
+};
 
-  switch (carouselType) {
-    case 'Project':
-      return ProjectImageCarousel;
-    case 'Discover':
-      return DiscoverCarousel;
-    default:
-      console.log('Error: Something went wrong, make sure carouselType is correct');
-      return ProjectImageCarousel;
-  }
+/**
+ * Primary carousel provider component that manages state and auto-scrolling.
+ * @param dataList array of React elements to display in the carousel
+ * @param children nested carousel components like CarouselButton, CarouselTabs, CarouselContent
+ * @returns A JSX Context Provider wrapping the carousel children
+ */
+export const Carousel = ({
+    dataList = [],
+    children,
+}: {
+    dataList?: ReactNode[];
+    children: ReactNode;
+}) => {
+    // Index of the currently active carousel slide
+    const [currentIndex, setCurrentIndex] = useState(0);
+    // Whether the user is hovering over the carousel (pauses auto-scroll)
+    const [hovering, setHovering] = useState(false);
+
+    // Boolean ref that temporarily disables auto-scroll after user interaction
+    // Prevents auto-scroll immediately after user clicks a button or hovers
+    const skipAuto = useRef(false);
+
+    /**
+     * Changes the current slide index, wrapping around if it exceeds bounds.
+     * Also sets skipAuto to true to prevent immediate auto-scroll after manual navigation.
+     *
+     * @param newIndex - Desired index to navigate to
+     */
+    const handleIndexChange = (newIndex: number) => {
+        skipAuto.current = true;
+        if (newIndex > dataList.length - 1) {
+            newIndex = 0;
+        } else if (newIndex < 0) {
+            newIndex = dataList.length - 1;
+        }
+
+        setCurrentIndex(newIndex);
+    };
+
+    /**
+     * Called when the user hovers over the carousel.
+     * @param hover - true if the mouse is over the carousel, false otherwise
+     */
+    const handleHover = (hover: boolean) => {
+        if (hover) {
+            skipAuto.current = true;
+            setHovering(true);
+        } else {
+            setHovering(false);
+        }
+    };
+
+    /**
+     * Auto-scroll logic: advances carousel by one slide unless skipAuto or hovering is true.
+     * Wraps around at the end of dataList.
+     */
+    const autoScroll = () => {
+        if (skipAuto.current) {
+            if (hovering) {
+                return;
+            } else {
+                skipAuto.current = false;
+                return;
+            }
+        }
+
+        if (currentIndex === dataList.length - 1) {
+            setCurrentIndex(0);
+        } else {
+            setCurrentIndex(currentIndex + 1);
+        }
+    };
+
+    // Interval effect to automatically scroll carousel every 10 seconds
+    // Stops scrolling when user interacts (hover or button click)
+    useEffect(() => {
+        const interval = setInterval(() => {
+            autoScroll();
+        }, 10_000);
+
+        return () => clearInterval(interval);
+    });
+
+    return (
+        <CarouselContext.Provider
+            value={{ currentIndex, handleIndexChange, handleHover, dataList }}
+        >
+            {children}
+        </CarouselContext.Provider>
+    );
 };
